Integrate premium backend, reports, plan tracking and reviewed plan flow

+++ lib/features/meals/domain/models/barcode_food_data.dart
@@ -6,9 +6,9 @@ class BarcodeFoodData {
   final String unit;
 
   final int calories;
-  final int protein;
-  final int fats;
-  final int carbs;
+  final double protein;
+  final double fats;
+  final double carbs;
 
   final String nutritionSource;
   final bool isEstimated;
